pyproject.toml
[project]
name = "backend-omnimind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=0.1.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "httpx>=0.27.0",
    "langchain-community>=0.4.1",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.5",
    "langsmith>=0.6.0",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.8.2",
    "python-multipart>=0.0.6",
    "supabase>=2.27.0",
    "typing-extensions>=4.8.0",
    "uvicorn>=0.40.0",
]
